[Feature] Add SwiftLint, divide some files # Update summary * Added SwiftLint support * Resolve some swiftlint errors and warnings * Divided some functions to extensions
