+++ Example/TableViewController.swift
@@ -41,7 +41,8 @@ class TableViewController: UITableViewController {
         return cell
     }
     
-    override func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+    override func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell,
+                            forRowAt indexPath: IndexPath) {
         guard let cell = cell as? MenuTableViewCell else { return }
         cell.dropdownMenu.reload()
     }
